Move pytest fixtures to standard conftest.py location
